% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-tree-sitter-test-package.R
\docType{package}
\name{test-package}
\alias{test}
\alias{test-package}
\title{test: Just A Test}
\description{
Just a test package.
}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
